Move src/containers/App.jsx => src/App.jsx
