+++ src/index.jsx
@@ -2,6 +2,6 @@ import React from 'react';
 import {render} from 'react-dom';
 
 // Main app
-import App from './containers/App.jsx';
+import App from './App.jsx';
 
 render(<App/>, document.getElementById('app'));
